compression/python/pytree/requirements.txt
immutabledict>=4.2.0
numpy>=1.26.4
orbax-checkpoint>=0.0.0

